pyproject.toml
[tool.poetry]
name = "urs"
version = "3.4.0"
description = "URS (Universal Reddit Scraper): A comprehensive Reddit scraping and OSINT command-line tool"
authors = ["Joseph Lai <urs_project@protonmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
colorama = "^0.4.6"
halo = "^0.0.31"
praw = "^7.7.0"
prettytable = "^3.7.0"
python-dotenv = "^1.0.0"
rich = "^13.3.5"
wordcloud = "^1.9.1.1"
pexpect = "^4.9.0"
celery = "^5.3.0"
redis = "^5.0.0"
boto3 = "^1.34.0"

[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
coverage = "^7.2.4"
isort = "^5.12.0"
maturin = "^0.14.17"
pytest = "^7.3.1"
pytest-cov = "^4.0.0"

[tool.maturin]
features = ["pyo3/extension-module"]
python-source = "urs"

[build-system]
requires = ["maturin>=0.14,<0.15", "poetry-core"]
build-backend = "maturin"
